Fixed a really big issue
Since the playTime() function was being called on every re-render, it would end up playing several of the same audio at the same time which is a big issue. It would also cause the speed and volume sliders to create and play a new track on top of the current one. All of these issues are now fixed